refactor: 서버 선택 ui 디렉토리 변경
- 헤더에 있지만 사실 꼭 헤더에 있을 필요는 없음
- 헤더의 다른 기능들과는 달리 이질적임
